Implement name inference for torch.dot (#24474)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/24474

torch.dot is a little weird. It ignores the names of its inputs to be
consistent with the rest of our matrix multiplication functions.

I've written the implementation using a helper function that is also
used by other matrix multiplication functions so that it is easy to
change the behavior.

Test Plan
- new tests [namedtensor ci]

Test Plan: Imported from OSS

Differential Revision: D16915802

Pulled By: zou3519

fbshipit-source-id: 628a6de1935357022cc92f4d23222736a70bb070