Only default USE_DISTRIBUTED=True on Linux (#25725)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/25725

After landing #25260 the macOS wheel builds started to fail. It turns
out that if not specified, the setup helpers default USE_DISTRIBUTED
to true on all platforms except Windows.

This commit updates that such that USE_DISTRIBUTED only defaults to
true on Linux. More work is needed to enable it by default on macOS.

[test wheel]

ghstack-source-id: 89571701

Test Plan: N/A

Differential Revision: D17211695

fbshipit-source-id: 185db2e3425e45e6b76bd09d70a84e57327ca20f