Add ivalue::type(), part 1 (#25439)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/25439

This introduces a type() method on IValue that returns the tagged type
of the IValue. The intention is that this value is always present/accurate,
making it possible for clients to recover the Type from an IValue.
Currently our APIs here are incomplete: they can sometimes recover a type but not always.

This PR adds the function, and cleans up remaining cases where Lists/Dicts are not
tagged. However, this information does not survive serialization unchanged.

A second PR will use the type information in the ClassType being serialized
to fixup the serialized ivalues to have the correct types again.
After this patch it will be save to remove our incomplete APIs for recovering types.

Test Plan: Imported from OSS

Differential Revision: D17125595

Pulled By: zdevito

fbshipit-source-id: 71c8c1a0e44762647e8f15f45d8ed73af8e6cb92