Refactor emitIsInstance (#26061)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/26061

This is in preparation for actually emitting a dynamic isinstance check instruction.
It re-arranges the  logic so that all the types and properties to check
against are in a flat list. In the future this flat list will be encoded
into an actual instruction if we determine that we cannot perform
the check statically.

Test Plan: Imported from OSS

Differential Revision: D17332062

Pulled By: zdevito

fbshipit-source-id: 4c0b65436f8e030170d469fe747e79de24bb24eb