Remove zero_dim_tensors only from _th_masked_fill_. (#28513)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/28513

This just ensures that (since we only have a Scalar implementation), if you pass in a Tensor that's not zero-dim you get a nice error message.

Instead of doing this with codegen, we do this in code at the ATen level.

Test Plan: Imported from OSS

Differential Revision: D18086969

Pulled By: gchanan

fbshipit-source-id: 83fe2c16046e243d573e033d033aa3844b03930a