Add an AutogradMeta factory. (#28593)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/28593

When I turn on Variable everywhere, I will need to be able to construct
AutogradMetas from TensorImpl.  But I cannot call the constructor directly
as it lives in another dynamic library. So I need another virtual factory interface
to be able to do this.

I also adjust the AutogradMeta constructor so that the TensorImpl argument is
optional. This argument is only needed if `requires_grad == True`, as we use it
to test if the variable is valid (only floating point tensors can have requires grad true).

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

Test Plan: Imported from OSS

Differential Revision: D18171161

Pulled By: ezyang

fbshipit-source-id: 3f2e86720899b3bda36ddd90244c2624645cc519